# Simplify array and multimap codecs