--- a/go/otel/oteltef/anyvaluearray.go
+++ b/go/otel/oteltef/anyvaluearray.go
@@ -28,6 +28,19 @@ func (e *AnyValueArray) init(parentModifiedFields *modifiedFields, parentModifie
 	e.parentModifiedBit = parentModifiedBit
 }
 
+// reset the array to its initial state, as if init() was just called.
+// Will not reset internal fields such as parentModifiedFields.
+func (e *AnyValueArray) reset() {
+	e.elems = e.elems[:0]
+}
+
+// fixParent sets the parentModifiedFields pointer to the supplied value.
+// This is used when the parent is moved in memory for example because the parent
+// an array element and the array was expanded.
+func (e *AnyValueArray) fixParent(parentModifiedFields *modifiedFields) {
+	e.parentModifiedFields = parentModifiedFields
+}
+
 // Clone() creates a deep copy of AnyValueArray
 func (e *AnyValueArray) Clone() AnyValueArray {
 	var clone AnyValueArray
@@ -60,10 +73,6 @@ func (e *AnyValueArray) markModified() {
 	e.parentModifiedFields.markModified(e.parentModifiedBit)
 }
 
-func (e *AnyValueArray) markUnmodified() {
-	e.parentModifiedFields.markUnmodified()
-}
-
 func (e *AnyValueArray) markModifiedRecursively() {
 	for i := 0; i < len(e.elems); i++ {
 		e.elems[i].markModifiedRecursively()
@@ -76,34 +85,6 @@ func (e *AnyValueArray) markUnmodifiedRecursively() {
 		e.elems[i].markUnmodifiedRecursively()
 	}
 
-}
-
-// markDiffModified marks fields in each element of this array modified if they differ from
-// the corresponding fields in v.
-func (e *AnyValueArray) markDiffModified(v *AnyValueArray) (modified bool) {
-	if len(e.elems) != len(v.elems) {
-		// Array lengths are different, so they are definitely different.
-		modified = true
-	}
-
-	// Scan the elements and mark them as modified if they are different.
-	minLen := min(len(e.elems), len(v.elems))
-	for i := 0; i < minLen; i++ {
-		if e.elems[i].markDiffModified(v.elems[i]) {
-			modified = true
-		}
-	}
-
-	// Mark the rest of the elements as modified.
-	for i := minLen; i < len(e.elems); i++ {
-		e.elems[i].markModifiedRecursively()
-	}
-
-	if modified {
-		e.markModified()
-	}
-
-	return modified
 }
 
 func copyAnyValueArray(dst *AnyValueArray, src *AnyValueArray) {
@@ -229,67 +210,6 @@ type AnyValueArrayEncoder struct {
 	elemEncoder *AnyValueEncoder
 	isRecursive bool
 	state       *WriterState
-	// lastValStack are last encoded values stacked by the level of recursion.
-	lastValStack AnyValueArrayEncoderLastValStack
-}
-type AnyValueArrayEncoderLastValStack []*AnyValueArrayEncoderLastValElem
-
-func (s *AnyValueArrayEncoderLastValStack) init() {
-	// We need one top-level element in the stack to store the last value initially.
-	s.addOnTop()
-}
-
-func (s *AnyValueArrayEncoderLastValStack) reset() {
-	// Reset all elements in the stack.
-	t := (*s)[:cap(*s)]
-	for i := 0; i < len(t); i++ {
-		t[i].reset()
-	}
-	// Reset the stack to have one element for top-level.
-	*s = (*s)[:1]
-}
-
-func (s *AnyValueArrayEncoderLastValStack) top() *AnyValueArrayEncoderLastValElem {
-	return (*s)[len(*s)-1]
-}
-
-func (s *AnyValueArrayEncoderLastValStack) addOnTopSlow() {
-	elem := &AnyValueArrayEncoderLastValElem{}
-	elem.init()
-	*s = append(*s, elem)
-	t := (*s)[0:cap(*s)]
-	for i := len(*s); i < len(t); i++ {
-		// Ensure that all elements in the stack are initialized.
-		t[i] = &AnyValueArrayEncoderLastValElem{}
-		t[i].init()
-	}
-}
-
-func (s *AnyValueArrayEncoderLastValStack) addOnTop() {
-	if len(*s) < cap(*s) {
-		*s = (*s)[:len(*s)+1]
-		return
-	}
-	s.addOnTopSlow()
-}
-
-func (s *AnyValueArrayEncoderLastValStack) removeFromTop() {
-	*s = (*s)[:len(*s)-1]
-}
-
-type AnyValueArrayEncoderLastValElem struct {
-	prevLen        int
-	elem           AnyValue
-	modifiedFields modifiedFields
-}
-
-func (e *AnyValueArrayEncoderLastValElem) init() {
-	e.elem.init(&e.modifiedFields, 1)
-}
-
-func (e *AnyValueArrayEncoderLastValElem) reset() {
-	e.elem = AnyValue{}
-	e.prevLen = 0
 }
 
 func (e *AnyValueArrayEncoder) Init(state *WriterState, columns *pkg.WriteColumnSet) error {
@@ -313,7 +233,6 @@ func (e *AnyValueArrayEncoder) Init(state *WriterState, columns *pkg.WriteColumn
 			return err
 		}
 	}
-	e.lastValStack.init()
 
 	return nil
 }
@@ -322,38 +241,18 @@ func (e *AnyValueArrayEncoder) Reset() {
 	if !e.isRecursive {
 		e.elemEncoder.Reset()
 	}
-
-	e.lastValStack.reset()
 }
 
 func (e *AnyValueArrayEncoder) Encode(arr *AnyValueArray) {
-	lastVal := e.lastValStack.top()
-	e.lastValStack.addOnTop()
-	defer func() { e.lastValStack.removeFromTop() }()
-
-	newLen := len(arr.elems)
 	oldBitLen := e.buf.BitCount()
 
-	lenDelta := newLen - lastVal.prevLen
-	lastVal.prevLen = newLen
+	// Write the length of the array.
+	newLen := len(arr.elems)
+	e.buf.WriteUvarintCompact(uint64(newLen))
 
-	e.buf.WriteVarintCompact(int64(lenDelta))
-
-	if newLen > 0 {
-		for i := 0; i < newLen; i++ {
-			if i == 0 {
-				// Compute and mark fields that are modified compared to the last encoded value.
-				arr.elems[i].markDiffModified(&lastVal.elem)
-			} else {
-				// Compute and mark fields that are modified compared to the previous element.
-				arr.elems[i].markDiffModified(arr.elems[i-1])
-			}
-
-			// Encode the element.
-			e.elemEncoder.Encode(arr.elems[i])
-		}
-		// Remember last encoded element.
-		copyAnyValue(&lastVal.elem, arr.elems[len(arr.elems)-1])
+	// Encode the elements of the array.
+	for i := 0; i < newLen; i++ {
+		e.elemEncoder.Encode(arr.elems[i])
 	}
 
 	// Account written bits in the limiter.
@@ -373,68 +272,6 @@ type AnyValueArrayDecoder struct {
 	column      *pkg.ReadableColumn
 	elemDecoder *AnyValueDecoder
 	isRecursive bool
-	// lastValStack are last decoded values stacked by the level of recursion.
-	lastValStack AnyValueArrayDecoderLastValStack
-}
-type AnyValueArrayDecoderLastValStack []*AnyValueArrayDecoderLastValElem
-
-func (s *AnyValueArrayDecoderLastValStack) init() {
-	// We need one top-level element in the stack to store the last value initially.
-	s.addOnTop()
-}
-
-func (s *AnyValueArrayDecoderLastValStack) reset() {
-	// Reset all elements in the stack.
-	t := (*s)[:cap(*s)]
-	for i := 0; i < len(t); i++ {
-		t[i].reset()
-	}
-	// Reset the stack to have one element for top-level.
-	*s = (*s)[:1]
-}
-
-func (s *AnyValueArrayDecoderLastValStack) top() *AnyValueArrayDecoderLastValElem {
-	return (*s)[len(*s)-1]
-}
-
-func (s *AnyValueArrayDecoderLastValStack) addOnTopSlow() {
-	elem := &AnyValueArrayDecoderLastValElem{}
-	elem.init()
-	*s = append(*s, elem)
-	t := (*s)[0:cap(*s)]
-	for i := len(*s); i < len(t); i++ {
-		// Ensure that all elements in the stack are initialized.
-		t[i] = &AnyValueArrayDecoderLastValElem{}
-		t[i].init()
-	}
-}
-
-func (s *AnyValueArrayDecoderLastValStack) addOnTop() {
-	if len(*s) < cap(*s) {
-		*s = (*s)[:len(*s)+1]
-		return
-	}
-	s.addOnTopSlow()
-}
-
-func (s *AnyValueArrayDecoderLastValStack) removeFromTop() {
-	*s = (*s)[:len(*s)-1]
-}
-
-type AnyValueArrayDecoderLastValElem struct {
-	prevLen     int
-	elem        *AnyValue
-	elemStorage AnyValue
-}
-
-func (e *AnyValueArrayDecoderLastValElem) init() {
-	e.elem = &e.elemStorage
-
-}
-
-func (e *AnyValueArrayDecoderLastValElem) reset() {
-	e.prevLen = 0
-	*e.elem = AnyValue{}
 }
 
 // Init is called once in the lifetime of the stream.
@@ -456,7 +293,6 @@ func (d *AnyValueArrayDecoder) Init(state *ReaderState, columns *pkg.ReadColumnS
 			return err
 		}
 	}
-	d.lastValStack.init()
 
 	return nil
 }
@@ -477,27 +313,18 @@ func (d *AnyValueArrayDecoder) Reset() {
 	if !d.isRecursive {
 		d.elemDecoder.Reset()
 	}
-	d.lastValStack.reset()
 }
 
 func (d *AnyValueArrayDecoder) Decode(dst *AnyValueArray) error {
-	lastVal := d.lastValStack.top()
-	d.lastValStack.addOnTop()
-	defer func() { d.lastValStack.removeFromTop() }()
-
-	lenDelta := d.buf.ReadVarintCompact()
-
-	newLen := lastVal.prevLen + int(lenDelta)
-	lastVal.prevLen = newLen
+	newLen := int(d.buf.ReadUvarintCompact())
 
 	dst.EnsureLen(newLen)
 
 	for i := 0; i < newLen; i++ {
-		err := d.elemDecoder.Decode(lastVal.elem)
+		err := d.elemDecoder.Decode(dst.elems[i])
 		if err != nil {
 			return err
 		}
-		copyAnyValue(dst.elems[i], lastVal.elem)
 	}
 
 	return nil