… benchmarks/testdata/astronomy-otelmetrics.stefz
… examples/jsonl/internal/jsonstef/jsonobject.go
… examples/jsonl/internal/jsonstef/jsonvalue.go
… examples/jsonl/internal/jsonstef/jsonvaluearray.go
… examples/jsonl/internal/jsonstef/modifiedfields.go
… examples/jsonl/internal/jsonstef/record.go
… examples/jsonl/json2stef.go
… examples/jsonl/makefile
… examples/profile/internal/profile/function.go
… examples/profile/internal/profile/labels.go
… examples/profile/internal/profile/labelvalue.go
… examples/profile/internal/profile/line.go
… examples/profile/internal/profile/linearray.go
… examples/profile/internal/profile/location.go
… examples/profile/internal/profile/locationarray.go
… examples/profile/internal/profile/mapping.go
… examples/profile/internal/profile/modifiedfields.go
… examples/profile/internal/profile/numvalue.go
… examples/profile/internal/profile/profilemetadata.go
… examples/profile/internal/profile/sample.go
… examples/profile/internal/profile/samplevalue.go
… examples/profile/internal/profile/samplevaluearray.go
… examples/profile/internal/profile/samplevaluetype.go
… examples/profile/internal/profile/stringarray.go
… examples/profile/makefile
… go/otel/manual_test.go
… go/otel/otel.stef
… go/otel/oteltef/envelope.go
… go/otel/oteltef/envelopeattributes.go
… go/otel/oteltef/event.go