## Summary

Important: this is a breaking format change!

This change significantly simplifies array and multimap codecs. It gets rid of the
complicated diff computations (especially for recursive data). Instead of attempting
to keep previous values and perform diff computations all diff logic is now much simpler
and is based on bookkeeping of plian "modified" flags which indicate a particular piece
of data is modified compared to that same piece of data in the last root record.

- Multimaps now track modification of each element's key and value individually, similarly
to how structs track modification of each field.
- Arrays now encode their lengths as is, instead of using delta of length.

All in all, the new approach is much simpler to implement and also cheaper computationally.

The change resulted in the following:

- Significant chunks of complicated code deleted. For example, eliminated last val stacks
  in encoders/decoders and got rid of markDiffModified familiy of funcs.
- Major performance improvements (about 2x) for array-heavy schemas such as Profile example.
- Minor performance improvements for other schemas (e.g. Otel schema).
- Very slight increase in encoded size for array-heavy schemas, virtually no size change for
  other schemas.

## How To Review

As always ignore the generated code (except to look at some samples to get a feel of it).
Look at the changes in *.tmpl files and in specification.md.

## Performance and Size Comparison Details

### Profile Schema

The deserialization performance of Profile schema is improved about 2x:


```
pkg: github.com/splunk/stef/examples/profile
cpu: Apple M2 Pro
                                                       │ bench_main.txt │         bench_current.txt          │
                                                       │     sec/op     │   sec/op     vs base               │
Deserialization/file=deser_stef.prof/format=stef-10        1227.1µ ± 3%   694.3µ ± 2%  -43.42% (p=0.002 n=6)
Deserialization/file=otelcol_otlp.prof/format=stef-10       237.6µ ± 1%   165.1µ ± 1%  -30.52% (p=0.002 n=6)
Deserialization/file=wsstreamasync.prof/format=stef-10      2.680m ± 1%   1.047m ± 1%  -60.91% (p=0.002 n=6)
geomean                                                     921.1µ        493.3µ       -46.44%

                                                       │ bench_main.txt │          bench_current.txt          │
                                                       │      B/op      │     B/op      vs base               │
Deserialization/file=deser_stef.prof/format=stef-10        1.444Mi ± 0%   1.081Mi ± 0%  -25.14% (p=0.002 n=6)
Deserialization/file=otelcol_otlp.prof/format=stef-10      442.3Ki ± 0%   385.1Ki ± 0%  -12.94% (p=0.002 n=6)
Deserialization/file=wsstreamasync.prof/format=stef-10     2.529Mi ± 0%   1.516Mi ± 0%  -40.07% (p=0.002 n=6)
geomean                                                    1.164Mi        871.3Ki       -26.90%

                                                       │ bench_main.txt │         bench_current.txt          │
                                                       │   allocs/op    │  allocs/op   vs base               │
Deserialization/file=deser_stef.prof/format=stef-10         19.77k ± 0%   11.93k ± 0%  -39.65% (p=0.002 n=6)
Deserialization/file=otelcol_otlp.prof/format=stef-10       4.827k ± 0%   3.598k ± 0%  -25.46% (p=0.002 n=6)
Deserialization/file=wsstreamasync.prof/format=stef-10      36.83k ± 0%   15.91k ± 0%  -56.78% (p=0.002 n=6)
geomean                                                     15.20k        8.808k       -42.07%
```

The size of encoded Profile data samples is very slightly increased, but big performance improvements
more than make up for it; it is a good tradeoff!

Before:
```
Uncompressed bytes sizes
File                 | pprof  | stef
deser_stef.prof      |  83343 |  83090 (0.997x)
otelcol_otlp.prof    |  31864 |  38769 (1.217x)
wsstreamasync.prof   |  92996 | 126985 (1.365x)
```

After:
```
Uncompressed bytes sizes
File                 | pprof  | stef
deser_stef.prof      |  83343 |  83768 (1.005x)
otelcol_otlp.prof    |  31864 |  38859 (1.220x)
wsstreamasync.prof   |  92996 | 128295 (1.380x)
```


### Otel Schema

The deserialization performance of Otel schema is also improved slightly:

```
pkg: github.com/splunk/stef/benchmarks
cpu: Apple M2 Pro
                               │ bench_main.txt │         bench_current.txt         │
                               │     sec/op     │   sec/op     vs base              │
SerializeNative/STEF/none-10        4.580m ± 2%   4.573m ± 1%       ~ (p=0.546 n=9)
DeserializeNative/STEF/none-10      1.699m ± 1%   1.627m ± 0%  -4.25% (p=0.000 n=9)
geomean                             2.789m        2.727m       -2.22%

                               │ bench_main.txt │         bench_current.txt         │
                               │   sec/point    │  sec/point   vs base              │
SerializeNative/STEF/none-10        68.50n ± 2%   68.39n ± 1%       ~ (p=0.546 n=9)
DeserializeNative/STEF/none-10      25.41n ± 1%   24.33n ± 0%  -4.25% (p=0.000 n=9)
geomean                             41.72n        40.79n       -2.23%

                               │ bench_main.txt │         bench_current.txt          │
                               │      B/op      │     B/op      vs base              │
SerializeNative/STEF/none-10       3.513Mi ± 0%   3.572Mi ± 0%  +1.68% (p=0.000 n=9)
DeserializeNative/STEF/none-10     842.3Ki ± 0%   899.7Ki ± 0%  +6.81% (p=0.000 n=9)
geomean                            1.700Mi        1.771Mi       +4.21%

                               │ bench_main.txt │         bench_current.txt          │
                               │   allocs/op    │  allocs/op   vs base               │
SerializeNative/STEF/none-10        2.913k ± 0%   3.079k ± 0%   +5.70% (p=0.000 n=9)
DeserializeNative/STEF/none-10      1.342k ± 0%   1.503k ± 0%  +12.00% (p=0.000 n=9)
geomean                             1.977k        2.151k        +8.80%
```

The size of Otel encoded data is virtually unchanged, with only astronomy-otelmetrics
dataset showing slight differences, where uncompressed size is slightly worse and 
compressed size is slightly better (so, overall a wash):

Before:
```
astronomy-otelmetrics.zst               Uncompressed           Zstd Compressed
                                     Bytes Ratio By/pt        Bytes Ratio By/pt
STEF                               7411568 19.68   9.4      1656025  3.81   2.1
```

After:
```
astronomy-otelmetrics.zst               Uncompressed           Zstd Compressed
                                     Bytes Ratio By/pt        Bytes Ratio By/pt
STEF                               7443972 19.59   9.5      1646140  3.83   2.1
```